requirements.txt
APScheduler==3.11.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face-recognition-models @ file:///home/pi/Downloads/face_recognition_models-0.3.0-py2.py3-none-any.whl
gpiozero
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pyudev==0.24.0
pyzbar==0.1.9
qrcode==8.2
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.0
responses==0.18.0
rfc3987==1.3.8
roman==3.3
rpi-lgpio==0.6
RPLCD==1.4.0
RTIMULib==7.2.1
Send2Trash==1.8.1b0
sense-hat==2.6.0
simplejpeg==1.8.1
ssh-import-id==5.10
# Core runtime dependencies
numpy==1.24.4
pandas==1.5.3
opencv-python==4.8.1.78          # For video & face processing
face_recognition==1.3.0          # Depends on dlib
dlib==19.24.2                    # Requires cmake, boost
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.3.0
requests==2.31.0
picamera2==0.3.15                # For capturing from PiCamera2 (libcamera-based)
RPLCD==1.3.1                     # For I2C LCD
RPi.GPIO==0.7.1
apscheduler==3.10.4

# Optional: for improved JPEG handling (large image safety)
imageio==2.34.0
